rename streak to transientspec
