+++ lumispy/signals/__init__.py
@@ -25,9 +25,9 @@ from .el_spectrum import ELSpectrum, LazyELSpectrum
 from .luminescence_transient import LumiTransient, LazyLumiTransient
 from .cl_transient import CLTransient, LazyCLTransient
 from .pl_transient import PLTransient, LazyPLTransient
-from .luminescence_streak import LumiTransientSpectrum, LazyLumiTransientSpectrum
-from .cl_streak import CLTransientSpectrum, LazyCLTransientSpectrum
-from .pl_streak import PLTransientSpectrum, LazyPLTransientSpectrum
+from .luminescence_transientspec import LumiTransientSpectrum, LazyLumiTransientSpectrum
+from .cl_transientspec import CLTransientSpectrum, LazyCLTransientSpectrum
+from .pl_transientspec import PLTransientSpectrum, LazyPLTransientSpectrum
 
 
 __all__ = [
