+++ lumispy/signals/luminescence_transientspec.py
